backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
firebase-admin==7.4.0
supabase==1.2.0
python-dotenv==1.2.2
roboflow==1.3.1
pillow==12.2.0
requests==2.33.1
gunicorn==23.0.0
